Moved keyboard-input to its own component
